requirements.txt
python-dotenv==1.1.1

# LangChain Ecosystem
langchain==1.2.8
langchain-core==1.2.8
langchain-ollama==1.0.1

# LangGraph
langgraph==1.0.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3

# Pydantic
pydantic==2.11.7
pydantic-settings==2.10.0

# Java parsing
javalang==0.13.0

# Additional dependencies
ollama==0.6.1
xxhash==3.5.0
rich==13.9.4
